refactor!: Use `double` for material

+++ Core/include/Acts/Material/AccumulatedMaterialSlab.hpp
@@ -46,7 +46,7 @@ class AccumulatedMaterialSlab {
   ///
   ///  Vacuum steps with a non-zero thickness can be added to account for holes
   ///  in material structures.
-  void accumulate(MaterialSlab slabAlongTrack, float pathCorrection = 1);
+  void accumulate(MaterialSlab slabAlongTrack, double pathCorrection = 1);
 
   /// Use the accumulated material to update the material variance
   ///
@@ -89,7 +89,7 @@ class AccumulatedMaterialSlab {
   /// If there have been additional calls to `.accumulate(...)` afterwards, the
   /// information is not part of the total average. The number of tracks is only
   /// updated on the call of `.trackAverage(...)`
-  std::pair<float, unsigned int> totalVariance() const;
+  std::pair<double, unsigned int> totalVariance() const;
 
  private:
   /// Averaged properties for a single track.
@@ -97,9 +97,9 @@ class AccumulatedMaterialSlab {
   /// Averaged properties over multiple tracks.
   MaterialSlab m_totalAverage;
   /// Averaged variance over multiple tracks.
-  float m_totalVariance = 0.0;
+  double m_totalVariance = 0;
   // Number of tracks contributing to the total average.
-  unsigned int m_totalCount = 0u;
+  unsigned int m_totalCount = 0;
 };
 
 }  // namespace Acts
